backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# Supabase
supabase>=2.0.0
python-dotenv>=1.0.0

# Auth
PyJWT>=2.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

